Made sure grayscale on/off worked, added args.task_name, removed all mentions of args.natural_language, added todos, turn_instruction_str_to_tensor to task_utils.py, apple and mug default, controller.stop() at end of other test cases without wrapper.
Added warning in comment about bad build path.
